Abstract out aws key
